requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
libsql-client>=0.3.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
greenlet>=3.0.0
alembic>=1.12.0

# Templates and static files
jinja2>=3.1.0
python-multipart>=0.0.6

# Authentication
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0

# Real-time features
sse-starlette>=1.6.0

# Data validation
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0